face_classification/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0
safetensors>=0.3.0
matplotlib>=3.5.0
pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
